refactor(#1257): keep server graphs out of server-fn files' client transform
The Start compiler strips handler/.server bodies from server fn files for
the client and DCEs their imports — but module-level and EXPORTED helpers
survive, shipping their whole server dependency graph to the browser in
dev (rolldown-vite dev does no cross-module tree-shaking). Only three
files actually leaked:

- functions/ai.ts: prepareBilling + streamScriptEnhancement +
  enhanceScriptToString (→ llm-client → @tanstack/ai adapter family +
  otel, ~10MB) → moved to lib/ai/script-enhancement.ts
- functions/smart-retry.ts: executeSmartRetry (→ fal-pricing-live +
  workflow client → #db-client → drizzle) → moved to
  lib/sequences/smart-retry.ts
- functions/model-assets.ts: createGeneratedAsset + validateAssetInput
  (→ workflow client + generation gate → #db-client) → moved to
  lib/models/generated-assets.ts

Server fn files now hold only schemas, types, and createServerFn chains;
handlers reference the lib implementations solely inside their bodies,
which the compiler strips. No dynamic imports needed — clients keep
calling server functions, and the server keeps static imports.
