+++ package.json
@@ -163,6 +163,8 @@
     "zod": "^4.4.3"
   },
   "devDependencies": {
+    "@babel/parser": "^8.0.4",
+    "@babel/types": "^8.0.4",
     "@clack/prompts": "^1.7.0",
     "@cloudflare/vite-plugin": "^1.52.1",
     "@copilotkit/aimock": "^1.38.0",
